api/services/postgres_client.py
# Placeholder for Neon Postgres client logic
import os
import psycopg2
from dotenv import load_dotenv

load_dotenv()

class PostgresService:
    def __init__(self):
        self.postgres_uri = os.getenv("POSTGRES_URI")
        self.conn = None
        # try:
        #     self.conn = psycopg2.connect(self.postgres_uri)
        # except psycopg2.OperationalError as e:
        #     print(f"Could not connect to Postgres: {e}")

    def get_metadata(self, doc_id):
        # This is where you would retrieve metadata for a document
        print(f"Getting metadata for doc_id: {doc_id} from Postgres...")
        # with self.conn.cursor() as cur:
        #     cur.execute("SELECT * FROM documents WHERE id = %s", (doc_id,))
        #     return cur.fetchone()
        return {"id": doc_id, "title": "Mock Title", "url": "/docs/intro"}

    def __del__(self):
        if self.conn:
            self.conn.close()

postgres_service = PostgresService()
